Add support for exclude and include lists of packages to search
When searching for externally installed (ie. system) packages,
skip packages that are listed in the 'exclude' list or include
only packages that are listed in the 'include' list when adding
'--system' as option to 'spack external find'.
Together with a list of packages it further limits the search
to the packages listed, in connection with the '-t' option,
only packages with the specified tags are considered. When using
with '--all' the list of all available packages is limited by
the 'exclude' or 'include' settings.
Users are able to maintain the 'exclude' and 'include' lists
in the 'system.yaml' config file:
This file has the following format:
  system:
     exclude:
       - package1
       - ..
or
  system:
     include:
       - package1
       - ..
For now, 'exclude' and 'include' lists are mutually exclusive.

Signed-off-by: Egbert Eich <eich@suse.com>